fixed issue with generating passes for langs with type variables
In the reification of a language, I used `TH.baseName`, which should give
a name without its module prefix. What I didn't realize is that it would
contain a lot of numerical garbage afterward: `ann` in the source would
become `ann_5146843573546845` or the like. Anyway, all I needed to do
was call an existing `fixup` function that I was already using to strip
the garbage off such typevars elsewhere.